updated gamelogic and moved popup message on client
