+++ Server/Assets/Scripts/GameLogic/EnemyInfo.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 9346694de056822429c892c8998240f7
+timeCreated: 1523590963
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
