+++ Server/Assets/Scripts/GameLogic/GameLogic.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 22026e9241e4a8c40b92eeade24472c6
+timeCreated: 1498417660
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
