+++ Server/Assets/Scripts/GameLogic/TowerInfo.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: b6c7aa1136f01ac4ea576e2182e29087
+timeCreated: 1523589416
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
